Solution to Check MSB of a Number

+++ Check MSB of a Number@@67882a731b4392d8fc1577e9/code.c
@@ -4,6 +4,6 @@ int main(){
     int n;
     scanf("%d",&n);
     int ans = 0;
-    ans = 1 << log2(n);
+    ans = 1 << (int)log2(n);
     printf("%d",ans);
 }
